[Feat] Make PreAuth into a separate route without x-state.
